Add Energy.Base.Binary class with byte manipulation and bitwise operation utilities
Implement comprehensive binary utility functions including byte order reversal for ushort and uint types, array comparison methods for byte/int/long/double/decimal arrays, endianness conversion functions (GetUInt32MSB/LSB), and bitwise operations (Not, Or, And, Xor, Rol, Ror) with support for ring buffer behavior and bit rotation.

+++ Energy.Core/Base/Byte.cs
@@ -28,14 +28,14 @@ namespace Energy.Base
         /// </summary>
         /// <param name="value">Unsigned int (32-bit)</param>
         /// <returns>Reversed numeric value</returns>
-        public static ulong Reverse(uint value)
+        public static uint Reverse(uint value)
         {
             byte b1 = (byte)((value >> 0) & 0xff);
             byte b2 = (byte)((value >> 8) & 0xff);
             byte b3 = (byte)((value >> 16) & 0xff);
             byte b4 = (byte)((value >> 24) & 0xff);
 
-            return (ulong)(b1 << 24 | b2 << 16 | b3 << 8 | b4 << 0);
+            return (uint)(b1 << 24 | b2 << 16 | b3 << 8 | b4 << 0);
         }
 
         #region Compare
